Move monitor tests into new test module
